fix: API Implement - Place and Tag Entity v4
Fix PatchMethod for Place (including Tag)
Fix PatchMethod for Reply (including Tag)
