+++ server/src/main/java/com/mainproject/server/dto/PageInfo.java
@@ -1,0 +1,13 @@
+package com.mainproject.server.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class PageInfo {
+    private int page;
+    private int size;
+    private long totalElements;
+    private int totalPages;
+}
