+++ server/src/main/java/com/mainproject/server/dto/SingleResponseDto.java
@@ -1,0 +1,10 @@
+package com.mainproject.server.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class SingleResponseDto<T> {
+    private T data;
+}
