+++ server/src/main/java/com/mainproject/server/place/dto/RequestFilterDto.java
@@ -1,0 +1,4 @@
+package com.mainproject.server.place.dto;
+
+public class RequestFilterDto {
+}
