+++ server/src/main/java/com/mainproject/server/place/dto/TagNameDto.java
@@ -1,0 +1,13 @@
+package com.mainproject.server.place.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class TagNameDto {
+
+    private String tagName;
+
+
+}
