+++ server/src/main/java/com/mainproject/server/reply/repository/ReplyRepository.java
@@ -1,0 +1,11 @@
+package com.mainproject.server.reply.repository;
+
+import com.mainproject.server.reply.entity.Reply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
+
+
+}
+
