--- server/src/main/java/com/mainproject/server/server/SwaggerConfig.java
@@ -1,40 +1,0 @@
-package com.mainproject.server.server;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-    @Bean
-    public Docket swaggerApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .useDefaultResponseMessages(false)
-                .groupName("team48")
-                .apiInfo(apiInfo())
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.mainproejct.server"))
-                .paths(PathSelectors.any())
-               //.paths(PathSelectors.ant("/api/**"))
-                .build();
-    }
-
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder().title("Team48")
-                .description("Team48 API Test for Swagger Documentation")
-                .termsOfServiceUrl("http://www.naver.com")
-                .license("Team48")
-                .licenseUrl("https://www.naver.com").version("1.0").build();
-    }
-
-}
