--- server/src/main/java/com/mainproject/server/server/dto/MultiResponseDto.java
@@ -1,18 +1,0 @@
-package com.mainproject.server.server.dto;
-
-import lombok.Getter;
-import org.springframework.data.domain.Page;
-
-import java.util.List;
-
-@Getter
-public class MultiResponseDto<T> {
-    private List<T> data;
-    private PageInfo pageInfo;
-
-    public MultiResponseDto(List<T> data, Page page) {
-        this.data = data;
-        this.pageInfo = new PageInfo(page.getNumber() + 1,
-                page.getSize(), page.getTotalElements(), page.getTotalPages());
-    }
-}
