--- server/src/main/java/com/mainproject/server/server/dto/SingleResponseDto.java
@@ -1,10 +1,0 @@
-package com.mainproject.server.server.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class SingleResponseDto<T> {
-    private T data;
-}
