--- server/src/main/java/com/mainproject/server/server/place/dto/RequestFilterDto.java
@@ -1,4 +1,0 @@
-package com.mainproject.server.server.place.dto;
-
-public class RequestFilterDto {
-}
