--- server/src/main/java/com/mainproject/server/server/place/dto/TagNameDto.java
@@ -1,13 +1,0 @@
-package com.mainproject.server.server.place.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class TagNameDto {
-
-    private String tagName;
-
-
-}
