--- server/src/main/java/com/mainproject/server/server/reply/repository/ReplyRepository.java
@@ -1,11 +1,0 @@
-package com.mainproject.server.server.reply.repository;
-
-import com.mainproject.server.server.reply.entity.Reply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ReplyRepository extends JpaRepository<Reply, Long> {
-
-
-}
-
