--- server/src/main/java/com/mainproject/server/server/user/entity/User.java
@@ -1,4 +1,0 @@
-package com.mainproject.server.server.user.entity;
-
-public class User {
-}
