+++ server/src/main/java/com/mainproject/server/user/entity/User.java
@@ -1,0 +1,4 @@
+package com.mainproject.server.user.entity;
+
+public class User {
+}
